+++ src/app/services/ontolex-lemon.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpManager } from "../utils/HttpManager";
+import { HttpManager, STRequestParams } from "../utils/HttpManager";
 import { Observable } from 'rxjs';
 import { IRI, AnnotatedValue } from '../models/Resources';
 import { map } from 'rxjs/operators';
@@ -16,7 +16,7 @@ export class OntoLexLemonServices {
      * Returns lexicons
      */
     getLexicons(): Observable<AnnotatedValue<IRI>[]> {
-        var params: any = {};
+        let params: STRequestParams = {};
         return this.httpMgr.doGet(this.serviceName, "getLexicons", params).pipe(
             map(stResp => {
                 return ResourceDeserializer.createIRIArray(stResp);
@@ -31,7 +31,7 @@ export class OntoLexLemonServices {
      * @param lexicon 
      */
     getLexicalEntriesByAlphabeticIndex(index: string, lexicon: IRI): Observable<AnnotatedValue<IRI>[]> {
-        var params: any = {
+        let params: STRequestParams = {
             index: index,
             lexicon: lexicon
         };
@@ -43,7 +43,7 @@ export class OntoLexLemonServices {
     }
 
     countLexicalEntriesByAlphabeticIndex(index: string, lexicon: IRI): Observable<number> {
-        var params: any = {
+        let params: STRequestParams = {
             index: index,
             lexicon: lexicon
         };
@@ -55,7 +55,7 @@ export class OntoLexLemonServices {
      * @param lexicalEntry 
      */
     getLexicalEntryLexicons(lexicalEntry: IRI): Observable<AnnotatedValue<IRI>[]> {
-        var params: any = {
+        let params: STRequestParams = {
             lexicalEntry: lexicalEntry
         };
         return this.httpMgr.doGet(this.serviceName, "getLexicalEntryLexicons", params).pipe(
@@ -63,6 +63,28 @@ export class OntoLexLemonServices {
                 return ResourceDeserializer.createIRIArray(stResp);
             })
         );
+    }
+
+    /**
+     * Returns the 2-digits index of the given lexicalEntry 
+     * @param lexicalEntry 
+     */
+     getLexicalEntryIndex(lexicalEntry: IRI): Observable<string> {
+        let params: STRequestParams = {
+            lexicalEntry: lexicalEntry
+        };
+        return this.httpMgr.doGet(this.serviceName, "getLexicalEntryIndex", params);
+    }
+
+    /**
+     * 
+     * @param lexicalEntry 
+     */
+    getLexicalEntryLanguage(lexicalEntry: IRI): Observable<string> {
+        let params: STRequestParams = {
+            lexicalEntry: lexicalEntry
+        };
+        return this.httpMgr.doGet(this.serviceName, "getLexicalEntryLanguage", params);
     }
 
 }